CBPS-984: collect logs after tools tests
- Set log_file option on all cbimport/cbexport tests
- Enable cbimport/cbexport logging by supplying --verbose alongside
--log-file
- Don't run cbbackupmgr collect-logs redundantly for cbimport/cbexport
tests
- Run cbbackupmgr collect-logs on all cbbackupmgr tests (most of them
were already doing it)

Reviewed-on: https://review.couchbase.org/c/perfrunner/+/169454
Tested-by: Build Bot <build@couchbase.com>

+++ scripts/upload_info.sh
@@ -1,6 +1,7 @@
 #!/bin/bash +x
 
 for FILE in *.zip; do
+    [ -f "$FILE" ] || continue
     env/bin/aws s3 cp --quiet ${FILE} s3://perf-artifacts/${BUILD_TAG}/${FILE}
     echo "https://s3-us-west-2.amazonaws.com/perf-artifacts/${BUILD_TAG}/${FILE}"
 done
